Upgrade pdf.js to v4.2.67 #62
* removed the method that atempts to automatically add the custom-mimetypes.xml file to the appropriate jar file
* code refactoring
