+++ macro-pdfviewer-api/src/main/resources/META-INF/components.txt
@@ -1,2 +1,2 @@
 com.xwiki.pdfviewer.internal.macro.PDFViewerMacro
-com.xwiki.pdfviewer.internal.filter.MJSMimeTypeRegistrar
+com.xwiki.pdfviewer.internal.MJSMimeTypeRegistrar
