--- macro-pdfviewer-api/src/main/resources/org/apache/tika/mime/custom-mimetypes.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<mime-info>
-  <mime-type type="text/javascript">
-    <glob pattern="*.mjs"/>
-    <sub-class-of type="text/plain"/>
-  </mime-type>
-</mime-info>
